# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{slug}": {
      "get": {
        "description": "Работает и для прежних коротких имён проекта: в ответе всегда действующее `slug`, и клиент заменяет им адрес в строке браузера (US-18). Не-участник получает 404 без каких-либо данных проекта.",
        "operationId": "ProjectsController_get",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "sladkiy-limit",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Проект по короткому имени",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "description": "Только администратор проекта. Переименование **не меняет** короткое имя в адресе: ранее отправленные ссылки продолжают работать (US-18).",
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "403": {
            "description": "Не администратор проекта (`project_forbidden`)"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Изменить название и описание",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "description": "Только администратор (US-17). Удаляются очереди, задачи, комментарии и вложения. Ключи очередей и прежние короткие имена остаются занятыми и повторно не выдаются (D-25, US-18).",
        "operationId": "ProjectsController_remove",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Проект удалён"
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "403": {
            "description": "Не администратор проекта (`project_forbidden`)"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Удалить проект",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/projects/{slug}/members": {
      "get": {
        "description": "Виден всем участникам проекта, включая читателя (US-14). Порядок: сначала администраторы, дальше по имени.",
        "operationId": "MembersController_list",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "sladkiy-limit",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор следующей страницы из поля `nextCursor`",
            "schema": {
              "minLength": 1,
              "maxLength": 256,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberListDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный курсор (`invalid_cursor`)"
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Участники проекта",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{slug}/members/{userId}": {
      "patch": {
        "description": "Только администратор (US-15). Администратор может понизить себя, если он не последний. Понижение последнего администратора — 409.",
        "operationId": "MembersController_changeRole",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "403": {
            "description": "Не администратор проекта (`project_forbidden`)"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          },
          "409": {
            "description": "Последний администратор проекта (`last_project_admin`)"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Изменить роль участника",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "description": "Исключить другого может только администратор; выйти самостоятельно может любой участник — для этого в `userId` передаётся собственный идентификатор (US-16). Последний администратор не уходит ни тем, ни другим путём — 409.\n\nЗадачи исключённого остаются, поле «Исполнитель» в них очищается, и на каждую такую задачу пишется запись истории в той же транзакции (D-31).",
        "operationId": "MembersController_remove",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveMemberResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "403": {
            "description": "Не администратор проекта (`project_forbidden`)"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          },
          "409": {
            "description": "Последний администратор проекта (`last_project_admin`)"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Исключить участника или выйти из проекта",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "SL Tracker API",
    "description": "HTTP API трекера задач SL Tracker. Единственный источник правды по контракту: файл генерируется из кода, руками не редактируется.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sl_session",
        "description": "Сессия веб-клиента: httpOnly Secure SameSite=Lax cookie (ADR-0001)."
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Сессия будущего мобильного клиента (ADR-0002)."
      }
    },
    "schemas": {
...
      "ProjectMemberPreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "Анна Иванова"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "enum": [
              "admin",
              "member",
              "reader"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "avatarUrl",
          "role"
        ]
      },
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "sladkiy-limit",
            "description": "Действующее короткое имя в адресе. Если проект запрошен по прежнему короткому имени, здесь всё равно действующее — клиенту следует заменить адрес в строке браузера на него (US-18)."
          },
          "name": {
            "type": "string",
            "example": "Сладкий лимит",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Markdown, до 5000 символов"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true,
            "description": "Подписанная ссылка на обложку со сроком жизни 10 минут. Бакет не публичный: ссылка без подписи содержимое не отдаёт. `null` — обложки нет, клиент рисует заглушку."
          },
          "role": {
            "enum": [
              "admin",
              "member",
              "reader"
            ],
            "type": "string",
            "description": "Роль текущего пользователя в этом проекте (permissions.md, 2.1)."
          },
          "memberCount": {
            "type": "number",
            "description": "Сколько всего участников в проекте"
          },
          "members": {
            "description": "Первые участники для группы аватаров: администраторы, затем по имени.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectMemberPreviewDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "coverUrl",
          "role",
          "memberCount",
          "members",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Новое название. Переименование **не меняет** короткое имя в адресе: ранее отправленные ссылки продолжают работать (US-18)."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 5000,
            "description": "`null` или пустая строка убирают описание."
          }
        }
      },
...
      "ProjectMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Идентификатор пользователя"
          },
          "displayName": {
            "type": "string",
            "example": "Анна Иванова"
          },
          "email": {
            "type": "string",
            "example": "anna@example.com"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "enum": [
              "admin",
              "member",
              "reader"
            ],
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Когда вступил в проект"
          },
          "isSelf": {
            "type": "boolean",
            "description": "Это текущий пользователь: в списке помечается «(вы)»"
          }
        },
        "required": [
          "userId",
          "displayName",
          "email",
          "avatarUrl",
          "role",
          "joinedAt",
          "isSelf"
        ]
      },
      "ProjectMemberListDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Сначала администраторы, дальше по имени (design/screens/project.md)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectMemberDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "number",
            "description": "Всего участников в проекте"
          }
        },
        "required": [
          "items",
          "nextCursor",
          "total"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "admin",
              "member",
              "reader"
            ],
            "type": "string",
            "description": "Новая роль. Понизить последнего администратора нельзя — 409 `last_project_admin` (permissions.md, п. 7)."
          }
        },
        "required": [
          "role"
        ]
      },
      "RemoveMemberResultDto": {
        "type": "object",
        "properties": {
          "unassignedIssues": {
            "type": "number",
            "description": "Сколько задач осталось без исполнителя: у задач исключённого поле «Исполнитель» очищается, и на каждую пишется запись истории (D-31).",
            "example": 3
          }
        },
        "required": [
          "unassignedIssues"
        ]
      },
...
    }
  }
}